try fix bazel undefined error
