+++ pkg/executor/select_test.go
@@ -15,15 +15,12 @@
 package executor_test
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/pingcap/tidb/pkg/domain"
 	"github.com/pingcap/tidb/pkg/executor"
 	"github.com/pingcap/tidb/pkg/parser/ast"
-	"github.com/pingcap/tidb/pkg/parser/mysql"
 	"github.com/pingcap/tidb/pkg/util/mock"
-	"github.com/stretchr/testify/require"
 )
 
 func BenchmarkResetContextOfStmt(b *testing.B) {
@@ -33,39 +30,4 @@ func BenchmarkResetContextOfStmt(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		executor.ResetContextOfStmt(ctx, stmt)
 	}
-}
-
-func TestImportIntoShouldHaveSameFlagsAsInsert(t *testing.T) {
-	insertStmt := &ast.InsertStmt{}
-	importStmt := &ast.ImportIntoStmt{}
-	insertCtx := mock.NewContext()
-	importCtx := mock.NewContext()
-	domain.BindDomain(insertCtx, &domain.Domain{})
-	domain.BindDomain(importCtx, &domain.Domain{})
-	for _, modeStr := range []string{
-		"",
-		"IGNORE_SPACE",
-		"STRICT_TRANS_TABLES",
-		"STRICT_ALL_TABLES",
-		"ALLOW_INVALID_DATES",
-		"NO_ZERO_IN_DATE",
-		"NO_ZERO_DATE",
-		"NO_ZERO_IN_DATE,STRICT_ALL_TABLES",
-		"NO_ZERO_DATE,STRICT_ALL_TABLES",
-		"NO_ZERO_IN_DATE,NO_ZERO_DATE,STRICT_ALL_TABLES",
-	} {
-		t.Run(fmt.Sprintf("mode %s", modeStr), func(t *testing.T) {
-			mode, err := mysql.GetSQLMode(modeStr)
-			require.NoError(t, err)
-			insertCtx.GetSessionVars().SQLMode = mode
-			require.NoError(t, executor.ResetContextOfStmt(insertCtx, insertStmt))
-			importCtx.GetSessionVars().SQLMode = mode
-			require.NoError(t, executor.ResetContextOfStmt(importCtx, importStmt))
-
-			insertTypeCtx := insertCtx.GetSessionVars().StmtCtx.TypeCtx()
-			importTypeCtx := importCtx.GetSessionVars().StmtCtx.TypeCtx()
-			require.EqualValues(t, insertTypeCtx.Flags(), importTypeCtx.Flags())
-		})
-	}
-
 }
